+++ WebApplication3/WebApplication3/Service/SupplierService.cs
@@ -27,6 +27,10 @@ namespace WebApplication3.Service
         /// <returns>id поставщика</returns>
         public async Task<int> AddAsync(Supplier obj, CancellationToken token)
         {
+            if (obj == null)
+            {
+                throw new ArgumentNullException();
+            }
             return await repository.AddAsync(obj, token);
         }
 
@@ -39,6 +43,10 @@ namespace WebApplication3.Service
         /// <returns>void</returns>
         public async Task DeleteAsync(int key , CancellationToken token)
         {
+            if (key <= 0)
+            {
+                throw new Exception("id не может быть 0");
+            }
             await repository.DeleteAsync(key, token);
         }
 
@@ -50,6 +58,10 @@ namespace WebApplication3.Service
         /// <returns>void</returns>
         public async Task UpdateAsync(Supplier obj, CancellationToken token)
         {
+            if (obj == null)
+            {
+                throw new ArgumentNullException();
+            }
             await repository.UpdateAsync(obj, token);
         }
 
@@ -71,6 +83,10 @@ namespace WebApplication3.Service
         /// <returns>Асинхронныя операция, которая возвращает атрибут сущности</returns>
         public async Task<Supplier> GetByIdAsync(int key, CancellationToken token)
         {
+            if (key <= 0)
+            {
+                throw new Exception("id не может быть 0");
+            }
             return await repository.GetByIdAsync(key, token);
         }
     }
